Mp3_audio_sink: add equalizer and volume control
Thirty-two equalizer bands may be adjusted using positive and negative
real numbers. The volume may also be be adjusted linearly using a
positive real number relative to 1.0. The volume adjustment is made in
the processing chain and the output is floating-point, so it is not
recommended to adjust the volume here but at the mixer.

<config>
	<volume linear="0.5"/> <!-- default, half volume -->
	<eq band="14" value="-0.2"/>
	<eq band="15" value="-0.1"/>
	<eq band="16" value="0.0"/> <!-- default, no adjustment -->
	<eq band="17" value="0.1"/>
	<eq band="18" value="0.2"/>
</config>