SSH terminal client
A client of the Terminal service that connects to a shell on a remote
SSH server.

It is configured with a "host" file found in its root directory with the
following format:
<host name="sdf-eu.org" port="22" user="root" pass="foo" known="yes"/>

The port, pass, and known attributes are optional. The client will first
try to authenticate with a keypair found in the root directory with a
fallback to password authentication. The client will automatically
disconnect from hosts that are not found in "/known_hosts", unless the
"known" attribute is set to the negative in the host file.