Port of cmus (git ~= 2.8.0-rc3)
“cmus is a small, fast and powerful console music player for Unix-like
operating systems (https://cmus.github.io/).”

Only the OSS output plugin is enabled and requires a properly
configured VFS to function.

For the time being only the libmad, e.g., MP3, input plugin is
configured in the pkg runtime as the other plugins might depend on
resampling or recoding to 44100/s16.

Fixes #232.