Nic_bus server
A minimal Nic service that switches ethernet packets between Nic
sessions. Packets are only switched between MAC addresses of directly
connected sessions, attaching a device driver and switching to
additional MACs on further ethernet segments is not supported. That is
to say that is component is neither a hub nor a switch.

This server is intended for creating private LANs between components and
supports arbitrary level 3 protocols (IPv4 and IPv6).