jdk: just reserve FDs in 'socketpair'
socketpair() is used in libnio/libnet to create "marker FDs", which are
only used as destination FD in dup2(). So it is safe to just reserve
those descriptors. Otherwise, calling functions may use uninitialized
stack variables after successful return of socketpair(), which results
in closing arbtrary fds later on.

Fixes #192