sd_card_drv: adapt omap4 and exynos5 to mainline
The sd_card_drv got changed to use the platform driver API
in genodelabs/genode#3956. This commit does not change the
drivers of exynos5 and omap4 to use it too, but at least keep
its functionality, and let it compile again.
To use the platform API, the specific platform drivers for
both SoC needs to be changed to use the modern platform API first.