Port sndio (1.5.0-git)
"Sndio is a small audio and MIDI framework part of the OpenBSD project
and ported to FreeBSD, Linux and NetBSD. It provides a lightweight audio
& MIDI server and a fully documented user-space API to access either the
server or directly the hardware in a uniform way. Sndio is designed to
work for desktop applications, but pays special attention to
synchronization mechanisms and reliability required by music
applications. Reliability through simplicity are part of the project
goals."

Fixes #263