jdk: upgrade RMA of VM area correctly
The VM area is a region map where JDK stores it's byte and
compiled code. When many dataspaces get attached the RM connection must
be upgraded (RAM and/or caps). Until now, the RAM update went to the parent,
and thus, the quota for the connection was not upgraded leading to an
infinitive update loop.

issue #311