seoul: fix invocation of isoinfo for UTF-8 systems
On my system, the invocation of the isoinfo tool fails because it
outputs "Setting input-charset to 'UTF-8' from locale." to stderr,
despite not returning an error code and successfully outputting the
requested grub.cfg file.

As a workaround, redirect stderr to /dev/null to make the run script
work on systems with non-C locale.

Fixes #314