scip: build still as c++17
as the ported versions of soplex and scip are quite old and can't be
built with c++20 we define for this two libraries c++17 as the to be
used standard.

This is of course just a workaround, the clean (but more labourios)
solution is to upgrade the libraries to a more current version.