seoul: improve cpu utilization
by replacing the global mutex (sync motherboard) with fine grained
locking in the models and in the Seoul app for Genode.

The commit improves the parallel build on all vCPUs appreciable.

Before this commit, the vCPU utilization did not exceed ~70%. With this
change the vCPU utilization is above 98%. The test case, of building vbox
for Genode inside the VM with 4 vCPUs, succeed 40s earlier on a T460p notebook
running Sculpt 23.10.