sdl2: use initial mode if window size is undefined
This commit is a follow-up to the recent mode info handling change
and restores the intended behavior by considering the initial mode
specified via the component's configuration when the window size is
undefined. This covers the use-case where the component is started
for the first time, no prior window-layouter rule is present and
thus the size should be limited to the configured value. In case
the component is started again afterwards the size is not undefined
any longer and the stored value, ignoring the initial mode, is used
instead.

Issue genodelabs/genode#5353.
Issue genodelabs/genode#5356.