sdl2: fix delivering key presses
This commit addresses a regression introduced with the preliminary
SDL_TEXTINPUT support that inhibits key press events (noticeable when
using WASD in games).

The 'SDL_IsTextInputActive()' check returns true when the textinput
subsystem is enabled. This is either the case when SDL2 calls
'SDL_StartTextInput()' itself (might happen during initialization)
or when it is enabled by an application (which can happen
unconditionally).

Issue genodelabs/genode#5356.